[FIX] Upgrade frontend application dependencies
[FIX] Fixed up the validate email address, models, file service and Spring Boot version

+++ src/main/kotlin/dev/joguenco/roqui/information/model/Information.kt
@@ -11,7 +11,7 @@ import org.hibernate.annotations.Immutable
 @Table(name = "v_ele_information")
 class Information {
 
-    @Id val id: Long? = null
+    @Id val id: Int? = null
 
     @Column(name = "identification") val identification: String? = null
